// powerapps-project-planner-20260612/src/generated/models/Vibe_budgetlinesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_budgetlinesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_budgetlinesstatecode = keyof typeof Vibe_budgetlinesstatecode;
export const Vibe_budgetlinesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_budgetlinesstatuscode = keyof typeof Vibe_budgetlinesstatuscode;
export const Vibe_budgetlinesvibe_budgetlinestatus = {
  100000000: 'PLANNED',
  100000001: 'APPROVED',
  100000002: 'SPENT'
} as const;
export type Vibe_budgetlinesvibe_budgetlinestatus = keyof typeof Vibe_budgetlinesvibe_budgetlinestatus;
export const Vibe_budgetlinesvibe_category = {
  100000000: 'LICENSES',
  100000001: 'INFRASTRUCTURE',
  100000002: 'TRAVEL',
  100000003: 'TRAINING',
  100000004: 'EXTERNAL_SERVICES',
  100000005: 'OTHER'
} as const;
export type Vibe_budgetlinesvibe_category = keyof typeof Vibe_budgetlinesvibe_category;

export interface Vibe_budgetlinesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_budgetlinesstatecode;
  statuscode?: Vibe_budgetlinesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_actualamount?: number;
  vibe_budgetlineid: string;
  vibe_budgetlinestatus?: Vibe_budgetlinesvibe_budgetlinestatus;
  vibe_category?: Vibe_budgetlinesvibe_category;
  vibe_description?: string;
  vibe_estimatedamount?: number;
  vibe_name?: string;
  "vibe_phaseid@odata.bind"?: string;
  "vibe_projectid@odata.bind"?: string;
}

export interface Vibe_budgetlines extends Vibe_budgetlinesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_budgetlinestatusname?: string;
  vibe_categoryname?: string;
  vibe_phaseidname?: string;
  vibe_projectidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_phaseid?: object;
  _vibe_phaseid_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
}