// powerapps-project-planner-20260612/src/generated/models/Vibe_invoicesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Vibe_invoicesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Vibe_invoicesstatecode = keyof typeof Vibe_invoicesstatecode;
export const Vibe_invoicesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Vibe_invoicesstatuscode = keyof typeof Vibe_invoicesstatuscode;

export interface Vibe_invoicesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Vibe_invoicesstatecode;
  statuscode?: Vibe_invoicesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  vibe_amount?: number;
  vibe_dateissued?: string;
  vibe_datepaid?: string;
  vibe_duedate?: string;
  vibe_invoiceid: string;
  vibe_invoicenumber?: string;
  vibe_notes?: string;
  "vibe_projectid@odata.bind"?: string;
}

export interface Vibe_invoices extends Vibe_invoicesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  vibe_projectidname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  vibe_projectid?: object;
  _vibe_projectid_value?: string;
}